Restructure package and update refs accordingly
