+++ docs/configuration.md
@@ -2,7 +2,7 @@
 Per default, the crawler will visit every qualified link it can find at retrieve the corresponding page content.
 Of course, you can customize and extend this behaviour.
 
-::: microwler.core.crawler.Crawler
+::: microwler.crawler.Microwler
     rendering:
       show_root_heading: true
 
@@ -10,7 +10,7 @@ Of course, you can customize and extend this behaviour.
 You can further tune your crawler using the `settings` parameter. It holds various configurations
 for the crawler itself, but also for any extensions/plugins you might want to use or build.
 
-> Note: You don't need to create a `Settings` object. Just put your settings in a `dict` as shown in the more complex [examples](/examples)
+> Note: You don't need to create a `Settings` object. Just put your settings in a `dict`.
 
 ### Crawler settings
 
@@ -43,6 +43,6 @@ The folder in which you want to save exported data files
 > Default: `${CWD}/projects`
 
 #### `settings.exporters`
-A list of export plugins inheriting from [microwler.core.export.BaseExporter][].
+A list of export plugins inheriting from [microwler.export.BaseExporter][].
 > Default: `[]`
