+++ docs/faq.md
@@ -9,7 +9,7 @@ server and the internet connection between them.
 Currently, there are two ways to do this:
 
 1. Use `crawler.data` to obtain a list of either data dictionaries if `selectors` are defined or the HTML documents
-2. Use `crawler.pages` to get the results as list of [microwler.core.scrape.Page][] objects
+2. Use `crawler.pages` to get the results as list of [microwler.scrape.Page][] objects
 3. Use `crawler.cache` to get the database as list of dictionaries
 
 #### What are *transformers*?
@@ -17,13 +17,11 @@ It sounds more complex than it really is: a *transformer* is any Python callable
 which works on a data dictionary. **Microwler** will inject every crawled page's `data`
 into a given transformer function in order to manipulate it after scraping, i.e. do some text processing.
 
-An example can be found [here](https://github.com/INNOVINATI/microwler/blob/master/test.py).
+An example can be found in the [advanced test case](https://github.com/INNOVINATI/microwler/blob/master/test_cases.py).
 
 #### Can I persist results, i.e. in a database?
 Yes. In fact, **Microwler** comes with a built-in caching system for storing results on-disk.
-You can enable this feature via the `caching` setting. 
-
-Per default caches will be stored in `${CWD}/.microwler/cache`. 
+You can enable this feature via the `caching` setting. Per default caches will be stored in `${CWD}/.microwler/cache`. 
 
 Optionally, you can activate *incremental crawling* using the `delta_crawl` setting.
