+++ docs/index.md
@@ -22,8 +22,11 @@ Optionally, you can also:
 - crawl dynamic pages (coming soon)
 
 <hr>
+<p>
 
-Copyright 2020-2021 Maximilian Wolf
+</p>
+<small>
+**Copyright 2020-2021 Maximilian Wolf**
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -36,3 +39,4 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
+</small>
