+++ microwler/cli/cmd.py
@@ -102,6 +102,7 @@ def clear_cache(ctx):
 
 
 @click.command('serve')
+@click.option('-h', '--host', type=str, default='localhost', help='The host to run the webservice on.')
 @click.option('-p', '--port', type=int, default=5000, help='The port to run the webservice on.')
 def start_server(port):
     """ Start the built-in webservice """
