+++ microwler/export.py
@@ -3,7 +3,7 @@ import logging
 import os
 from datetime import datetime
 
-from microwler.core.settings import Settings
+from microwler.settings import Settings
 
 LOG = logging.getLogger(__name__)
 
@@ -11,7 +11,7 @@ LOG = logging.getLogger(__name__)
 class BaseExporter:
     """
     Use this class to build your custom export functionality, i.e. send data per HTTP or SMTP.
-    The Crawler instance will call `export()` once it's done with everything else.
+    The Microwler instance will call `export()` once it's done with everything else.
     You can plug your Exporter into the crawler by adding the class to `settings['exporters']`
     """
 
@@ -31,15 +31,16 @@ class BaseExporter:
     def export(self):
         """
         Export data to target destination
-        > Will be called by [microwler.crawler.Crawler][]
+        > Will be called by [microwler.crawler.Microwler][]
         """
         raise NotImplementedError()
 
 
 class FileExporter(BaseExporter):
     """
-    Take a look at [microwler.export](https://github.com/INNOVINATI/microwler/blob/master/microwler/export.py)
-    for examples on how to use this Exporter
+    This exporter will save data to your local filesystem. It currently provides
+    exports to JSON, CSV or HTML tables. Take a look at the following exporters
+    and their implementation to understand its usage.
     """
     extension = ''
 
@@ -65,6 +66,7 @@ class FileExporter(BaseExporter):
 
 
 class JSONExporter(FileExporter):
+    """ Exports to JSON files """
     extension = 'json'
 
     def convert(self):
@@ -73,6 +75,7 @@ class JSONExporter(FileExporter):
 
 
 class CSVExporter(FileExporter):
+    """ Exports to CSV files """
     extension = 'csv'
 
     def convert(self):
@@ -83,6 +86,7 @@ class CSVExporter(FileExporter):
 
 
 class HTMLExporter(FileExporter):
+    """ Exports data as <table> to HTML files """
     extension = 'html'
 
     def convert(self):
